deid/requirements.txt
# text
nltk==3.4
# progress bars
tqdm==4.32.1
# numbers
numpy==1.16.3
pandas==0.24.2
sympy==1.4
# modelling
torch==1.1.0
pytorch-transformers==1.2.0
seqeval==0.0.12
# testing
pytest==4.2.0